Replace toStdString.c_str with qPrintable | Code intention cleanup
toStdString.c_str / toStdString crash:
This is either caused by a compiler bug and / or problems with the STL
and / or Problems with different compiler settings for the QT DLL and
lux.

STL is header only and slightly different compile options for different compiles
that interact might result in different than expected parameters passed or data
layout in memory.

Using the QT supplied qPrintable macro does at this time seem to be
the best option to avoid these problems alltogether.


[1] http://developer.qt.nokia.com/forums/viewthread/1690
[2] http://bugreports.qt.nokia.com/browse/QTBUG-11445