Changed how external files are transmitted to slaves, see issue #614.
The master will hash each sending file and send the hash for each file to the slave. The slave then sees if it already has the file on disk, otherwise it requests the file from the master.

When external files to the slave, the files are streamed directly from disk on demand. The slave will hash the received data as it's streamed to disk, in order to verify successful transmission.

The slave uses the hash as part of the filename, so that received files can be re-used between sessions. This initial implementation does not add any code for managing or cleaning up the cached files.