Move Sample random number shape inside Sampler class
The nXD fields in the Sample class was the same for every Sample, so instead of
storing it inside the Sample (which leads to lot of duplicate an memory usage),
we now stores it inside the Sampler.

This means that Sample shape should now only be declared once for the Sampler
and not as many time as Sample instances.

This may improve memory usage in hybrid bidir and potentially leads to better
performances.

(see forum discussion: http://www.luxrender.net/forum/viewtopic.php?p=82031 )

Please review, it is mainly search and replace except in the renderers where
the call to RequestSamples must be moved.