Added "cooldown" time to metropolis sampler, helps convergence with non-default large mutation probability.
Certain scenes seem to converge extremely slow when running the Metropolis sampler with very high or very low large mutation probability. This can be due to a bug, or the effect of start-up bias. This patch addresses the issue by applying large mutation probability of 0.5 for a finite time window at the start of the render. We call this "cooldown time". It can also enabled/disabled with a parameter in the scene file (currently disabled on default).

Example:
Sampler "metropolis"
	"float largemutationprob" [0.01]
	"bool usecooldown" ["true"]

Please give it some testing. Defaulting it later on would improve the robustness of the algorithm, so you don't have to re-render on "poor" choice of plarge.

Thanks to Lord Crc for helping with it.