Start local rendering before connecting to slaves, see issue #1247.
Slaves are added in a separate thread after parsing has started, so that the core connects to the slaves during parsing.

The WorldEnd command will cause the master to perform an explicit flush, ensuring the scene is sent to all the slaves.

Any slaves added after WorldEnd has been reached will be updated individually.