Added ability to force a server to reset to idle state, see issue #1256
Introduced two new command line parameters for luxconsole:
"password" sets the server reset password, which is blank by default.
"resetserver" will send the reset command to the specified servers, using the reset password if supplied, and exits.

The reset password is authenticated using salted hashes in a challenge/response fashion.

Resetting an idle server has no effect.

If the slave is currently in the list of known servers, the master will try to disconnect the slave first and in any case remove it from the list.

If the slave is rendering when it receives the reset command, it will save the current progress to a resume film with the prefix "server_reset" before ending the session.

In the GUI a new "Reset" button has been added instead of using command line parameters. It will prompt for a password, which can be left blank if no password was set on the server.