Apply previous dispersion fix to Bidir.
Each bidir path stores the dispersion state inside a *single* flag in
BidirVertex, this patch is mainly the steps necessary to store this flag.

The main issue comes from the EvalPath function, which may modifie the flag for
future paths (which may not be dispersed). So I introduced a context manager
(using scope behavior in the class ContextSingle) which is able to restore the
single flag of the sw struct.

TODO: apply the same scheme to the GPU bidir.