Fix to build with boost 1.53 and Qt.
Some complexes macro in boost 1.53 are not compatible with Qt MOC (bug appears
to be in the MOC). This workaround deactivates the MOC for boost headers (which
by definition does not need to be pre-processed by Qt)

See the discussion on the arch mailing list.

Original patch by *Lukas Jirkovsky*

https://mailman.archlinux.org/pipermail/arch-dev-public/2013-March/024720.html